--- equipment_management_app/src/header.tsx
@@ -1,21 +1,0 @@
-function header() {
-  return (
-    <header>
-      <h2>備品管理システム(仮)</h2>
-      <img src="./images/loan_history.png" alt="貸出履歴管理のicon" />
-      <p>貸出履歴管理</p>
-      <img src="./images/equipment_management.png" alt="備品管理のicon" />
-      <p>備品管理</p>
-      <img src="./images/student_management.png" alt="学生管理のicon" />
-      <p>学生管理</p>
-      <img src="./images/teacher_management.png" alt="教員管理のicon" />
-      <p>教員管理</p>
-      <img src="./images/inventory.png" alt="棚卸のicon" />
-      <p>棚卸</p>
-      <img src="./images/category_management.png" alt="カテゴリー管理のicon" />
-      <p>カテゴリー管理</p>
-      <img src="./images/logout.png" alt="ログアウトのicon" />
-      <p>ログアウト</p>
-    </header>
-  );
-}
